+++ app/controllers/sessions_controller.rb
@@ -6,9 +6,14 @@ class SessionsController < ApplicationController
   def create
     user = User.find_by(email: params[:email])
     if user && user.authenticate(params[:password])
-      session[:user_id] = user.id
-      flash[:success] = 'You are signed in, enjoy!'
-      redirect_to home_path
+      if user.active?
+        session[:user_id] = user.id
+        flash[:success] = 'You are signed in, enjoy!'
+        redirect_to home_path
+      else
+        flash[:danger] = "Your account has been suspended, please contact customer service."
+        redirect_to sign_in_path
+      end
     else
       flash[:danger] = "Invalid email or password."
       redirect_to sign_in_path
