+++ app/controllers/stripe_event_handler_controller.rb
@@ -1,12 +1,9 @@
-class StripeHandler::ChargesController < ApplicationController
+class StripeEventHandlerController < ApplicationController
   skip_before_action :verify_authenticity_token, only: :create
 
   def create
     event = Stripe::Event.retrieve(params[:id])
-    if event.type == 'charge.succeeded'
-      ChargeSuccess.new(event: event).process
-    end
-
+    StripeEventHandler.process(event: event)
     render nothing: true, status: 201
   rescue Stripe::APIConnectionError, Stripe::StripeError
     render nothing: true, status: 400
