+++ app/models/user.rb
@@ -32,4 +32,8 @@ class User < ActiveRecord::Base
   def can_follow?(another_user)
     !(self.follows?(another_user) || self == another_user)
   end
+
+  def deactivate!
+    update_column(:active, false)
+  end
 end
