+++ db/migrate/20180301132307_add_active_to_users.rb
@@ -1,0 +1,5 @@
+class AddActiveToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :active, :boolean, default: true
+  end
+end
