+++ spec/cassettes/StripeEventHandlerController/POST_create/charge_succeeded/creates_a_payment_with_the_webhook_from_stripe_for_charge_succeeded.yml
@@ -30,7 +30,7 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Wed, 28 Feb 2018 21:17:58 GMT
+      - Thu, 01 Mar 2018 14:10:00 GMT
       Content-Type:
       - application/json
       Content-Length:
@@ -50,7 +50,7 @@ http_interactions:
       Cache-Control:
       - no-cache, no-store
       Request-Id:
-      - req_JbAjUKsGc6ngPk
+      - req_Fhy6H8h2SFZ2c1
       Stripe-Version:
       - '2018-02-06'
       Strict-Transport-Security:
@@ -147,5 +147,5 @@ http_interactions:
           "type": "charge.succeeded"
         }
     http_version: 
-  recorded_at: Wed, 28 Feb 2018 21:17:58 GMT
+  recorded_at: Thu, 01 Mar 2018 14:10:00 GMT
 recorded_with: VCR 2.9.3
