+++ spec/cassettes/StripeEventHandlerController/POST_create/charge_succeeded/creates_the_payment_with_reference_id.yml
@@ -1,0 +1,151 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://api.stripe.com/v1/events/evt_1C0WXREB8VdfobgqqYDjFG3g
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - Stripe/v1 RubyBindings/3.9.2
+      Authorization:
+      - Bearer sk_test_fiAG8z1KPD3ixkm7nEMDheTD
+      Content-Type:
+      - application/x-www-form-urlencoded
+      X-Stripe-Client-User-Agent:
+      - '{"bindings_version":"3.9.2","lang":"ruby","lang_version":"2.2.9 p480 (2017-12-15)","platform":"x86_64-linux","engine":"ruby","publisher":"stripe","uname":"Linux
+        version 4.4.0-116-generic (buildd@lgw01-amd64-021) (gcc version 5.4.0 20160609
+        (Ubuntu 5.4.0-6ubuntu1~16.04.9) ) #140-Ubuntu SMP Mon Feb 12 21:23:04 UTC
+        2018","hostname":"BarkBark"}'
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Server:
+      - nginx
+      Date:
+      - Thu, 01 Mar 2018 14:10:01 GMT
+      Content-Type:
+      - application/json
+      Content-Length:
+      - '2413'
+      Connection:
+      - keep-alive
+      Access-Control-Allow-Credentials:
+      - 'true'
+      Access-Control-Allow-Methods:
+      - GET, POST, HEAD, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Expose-Headers:
+      - Request-Id, Stripe-Manage-Version, X-Stripe-External-Auth-Required, X-Stripe-Privileged-Session-Required
+      Access-Control-Max-Age:
+      - '300'
+      Cache-Control:
+      - no-cache, no-store
+      Request-Id:
+      - req_sXQlbVU5C54yTJ
+      Stripe-Version:
+      - '2018-02-06'
+      Strict-Transport-Security:
+      - max-age=31556926; includeSubDomains; preload
+    body:
+      encoding: UTF-8
+      string: |
+        {
+          "id": "evt_1C0WXREB8VdfobgqqYDjFG3g",
+          "object": "event",
+          "api_version": "2018-02-06",
+          "created": 1519830493,
+          "data": {
+            "object": {
+              "id": "ch_1C0WXQEB8VdfobgqEZBzQXe7",
+              "object": "charge",
+              "amount": 999,
+              "amount_refunded": 0,
+              "application": null,
+              "application_fee": null,
+              "balance_transaction": "txn_1C0WXREB8VdfobgqGeeASm34",
+              "captured": true,
+              "created": 1519830492,
+              "currency": "usd",
+              "customer": "cus_CPLDOcpgZ0fgQS",
+              "description": null,
+              "destination": null,
+              "dispute": null,
+              "failure_code": null,
+              "failure_message": null,
+              "fraud_details": {},
+              "invoice": "in_1C0WXQEB8VdfobgqGy2NfNy6",
+              "livemode": false,
+              "metadata": {},
+              "on_behalf_of": null,
+              "order": null,
+              "outcome": {
+                "network_status": "approved_by_network",
+                "reason": null,
+                "risk_level": "normal",
+                "seller_message": "Payment complete.",
+                "type": "authorized"
+              },
+              "paid": true,
+              "receipt_email": null,
+              "receipt_number": null,
+              "refunded": false,
+              "refunds": {
+                "object": "list",
+                "data": [],
+                "has_more": false,
+                "total_count": 0,
+                "url": "/v1/charges/ch_1C0WXQEB8VdfobgqEZBzQXe7/refunds"
+              },
+              "review": null,
+              "shipping": null,
+              "source": {
+                "id": "card_1C0WXPEB8VdfobgqjCQ7WJsd",
+                "object": "card",
+                "address_city": null,
+                "address_country": null,
+                "address_line1": null,
+                "address_line1_check": null,
+                "address_line2": null,
+                "address_state": null,
+                "address_zip": null,
+                "address_zip_check": null,
+                "brand": "Visa",
+                "country": "US",
+                "customer": "cus_CPLDOcpgZ0fgQS",
+                "cvc_check": "pass",
+                "dynamic_last4": null,
+                "exp_month": 11,
+                "exp_year": 2019,
+                "fingerprint": "pel6e9vrF4SvsbpN",
+                "funding": "credit",
+                "last4": "4242",
+                "metadata": {},
+                "name": "alexexample@example.com",
+                "tokenization_method": null
+              },
+              "source_transfer": null,
+              "statement_descriptor": null,
+              "status": "succeeded",
+              "transfer_group": null
+            }
+          },
+          "livemode": false,
+          "pending_webhooks": 0,
+          "request": {
+            "id": "req_Nds7EN1XnCgsqk",
+            "idempotency_key": null
+          },
+          "type": "charge.succeeded"
+        }
+    http_version: 
+  recorded_at: Thu, 01 Mar 2018 14:10:01 GMT
+recorded_with: VCR 2.9.3
